requirements.txt
flask
flask_restful
flask_cors
openai
python-dotenv